+++ ResumableFunctions.Handler/Core/CallPusher.cs
@@ -32,7 +32,7 @@ namespace ResumableFunctions.Handler.Core
             _methodIdsRepo = methodIdsRepo;
         }
 
-        public async Task<long> PushCall(PushedCall pushedCall)
+        public async Task<int> PushCall(PushedCall pushedCall)
         {
             _pushedCallsRepo.Add(pushedCall);
             await _context.SaveChangesAsync();
@@ -40,7 +40,7 @@ namespace ResumableFunctions.Handler.Core
             return pushedCall.Id;
         }
         static readonly SemaphoreSlim SemaphoreSlim = new(1, 1);
-        public async Task<long> PushExternalCall(PushedCall pushedCall, string serviceName)
+        public async Task<int> PushExternalCall(PushedCall pushedCall, string serviceName)
         {
             await SemaphoreSlim.WaitAsync();
             try
