+++ ResumableFunctions.Handler/InOuts/MethodsGroup.cs
@@ -2,13 +2,13 @@
 
 public class MethodsGroup : IEntity
 {
-    public long Id { get; internal set; }
+    public int Id { get; internal set; }
     public string MethodGroupUrn { get; internal set; }
     public List<WaitMethodIdentifier> WaitMethodIdentifiers { get; internal set; } = new();
     public List<MethodWait> WaitRequestsForGroup { get; internal set; }
 
     public DateTime Created { get; internal set; }
-    public long? ServiceId { get; set; }
+    public int? ServiceId { get; set; }
     public List<WaitTemplate> WaitTemplates { get; internal set; }
 }
